Support for legacy Windows systems has been returned (see description or comment of this commit)
To support legacy builds of Windows 10 (1803 and earlier), Windows 8.1, Windows 8, Windows 7, the ability to use curl ("curl.exe") has been added if it is located in the GRU folder. If curl is installed in any other way, you can use it by creating a file "curl.txt" in the GRU folder and writing the full path to the installed "curl.exe" in it. These methods are introduced as test methods (although they may not be).